requirements.txt
# ── ORACLE ──
oracledb>=2.0

# ── DATA ──
pandas>=2.0
faker>=24.0

# ── LLM ──
openai>=1.0
litellm>=1.0
anthropic>=0.25.0

# ── AGENTS ──
langchain>=0.2.0
langgraph>=0.1.0
langchain-openai>=0.1.0
langsmith>=0.1.0

# ── VALIDATION ──
pydantic>=2.0

# ── UTILITIES ──
python-dotenv>=1.0
requests>=2.31.0
httpx>=0.27.0

# ── DEV TOOLS ──
pytest>=8.0
pytest-asyncio>=0.23.0
black>=24.0
flake8>=7.0
mypy>=1.8.0
pre-commit>=3.7.0
